cycle10/pyproject.toml
[project]
name = "cycle10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==2.3.3",
    "pymysql==1.1.0",
    "python-dotenv==1.0.0",
    "werkzeug==2.3.7",
]
